main menu, options menu, and volume slider
Very bare bones, but can be built on. added an audio mixer to allow volume to be controlled, and made a new scene for the menu. all three buttons work (for now). chose fun colors for the text for it to match, but im guessing art team will probablt change that.